# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/activate/product-feature/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductFeature"
                ],
                "summary": "添加产品功能",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "添加产品功能参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddProductFeature"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加产品功能",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/product-feature/del": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductFeature"
                ],
                "summary": "删除产品功能",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "功能ID",
                        "name": "feature_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除产品功能",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/product-feature/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductFeature"
                ],
                "summary": "获取产品功能列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取产品功能列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddProductFeature": {
            "type": "object",
            "required": [
                "feature_code",
                "feature_name",
                "product_id"
            ],
            "properties": {
                "feature_code": {
                    "description": "功能编码",
                    "type": "string"
                },
                "feature_name": {
                    "description": "功能名称",
                    "type": "string"
                },
                "product_id": {
                    "description": "产品ID",
                    "type": "integer"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}